+++ unity/.gitignore
@@ -15,4 +15,9 @@
 /native_src/quickjs*
 /native_src/puerts-cpp-tester
 /native_src/build*
+/native_src_il2cpp/v8*
+/native_src_il2cpp/nodejs*
+/native_src_il2cpp/quickjs*
+/native_src_il2cpp/puerts-cpp-tester
+/native_src_il2cpp/build*
 /node_modules
