+++ unity/native_src_il2cpp/cmake/nodejs_16/backend.json
@@ -1,0 +1,82 @@
+{
+    "definition": ["WITH_NODEJS"],
+    "include": [
+        "/include",
+        "/deps/v8/include",
+        "/deps/uv/include"
+    ],
+    "skip": { "win": { "ia32": true }, "android": { "x64": true } },
+    "link": {
+        "win": {
+            "ia32": ["/Lib/Win32/libnode.lib"],
+            "x64": ["/Lib/Win64/libnode.lib"]
+        },
+        "android": {
+            "arm64": [
+                "/lib/Android/arm64-v8a/libbrotli.a",
+                "/lib/Android/arm64-v8a/libnode_stub.a",
+                "/lib/Android/arm64-v8a/libv8_libbase.a",
+                "/lib/Android/arm64-v8a/libcares.a",
+                "/lib/Android/arm64-v8a/libopenssl.a",
+                "/lib/Android/arm64-v8a/libv8_libplatform.a",
+                "/lib/Android/arm64-v8a/libhistogram.a",
+                "/lib/Android/arm64-v8a/libuv.a",
+                "/lib/Android/arm64-v8a/libv8_snapshot.a",
+                "/lib/Android/arm64-v8a/libllhttp.a",
+                "/lib/Android/arm64-v8a/libuvwasi.a",
+                "/lib/Android/arm64-v8a/libv8_zlib.a",
+                "/lib/Android/arm64-v8a/libnghttp2.a",
+                "/lib/Android/arm64-v8a/libv8_base_without_compiler.a",
+                "/lib/Android/arm64-v8a/libzlib.a",
+                "/lib/Android/arm64-v8a/libnode.a",
+                "/lib/Android/arm64-v8a/libv8_compiler.a",
+                "/lib/Android/arm64-v8a/libv8_initializers.a"
+            ],
+            "armv7": [
+                "/lib/Android/armeabi-v7a/libbrotli.a",
+                "/lib/Android/armeabi-v7a/libnode_stub.a",
+                "/lib/Android/armeabi-v7a/libv8_libbase.a",
+                "/lib/Android/armeabi-v7a/libcares.a",
+                "/lib/Android/armeabi-v7a/libopenssl.a",
+                "/lib/Android/armeabi-v7a/libv8_libplatform.a",
+                "/lib/Android/armeabi-v7a/libhistogram.a",
+                "/lib/Android/armeabi-v7a/libuv.a",
+                "/lib/Android/armeabi-v7a/libv8_snapshot.a",
+                "/lib/Android/armeabi-v7a/libllhttp.a",
+                "/lib/Android/armeabi-v7a/libuvwasi.a",
+                "/lib/Android/armeabi-v7a/libv8_zlib.a",
+                "/lib/Android/armeabi-v7a/libnghttp2.a",
+                "/lib/Android/armeabi-v7a/libv8_base_without_compiler.a",
+                "/lib/Android/armeabi-v7a/libzlib.a",
+                "/lib/Android/armeabi-v7a/libnode.a",
+                "/lib/Android/armeabi-v7a/libv8_compiler.a",
+                "/lib/Android/armeabi-v7a/libv8_initializers.a"
+            ]
+        },
+        "osx": {
+            "x64": ["/lib/macOS/libnode.93.dylib"],
+            "arm64": ["/lib/macOS_arm64/libnode.93.dylib"]
+        },
+        "linux": {
+            "x64": ["/lib/Linux/libnode.so.93"]
+        }
+    },
+    "copy": {
+        "win": {
+            "ia32": ["/lib/Win32/libnode.dll"],
+            "x64": ["/lib/Win64/libnode.dll"]
+        },
+        "android": {
+        },
+        "osx": {
+            "x64": ["/lib/macOS/libnode.93.dylib"],
+            "arm64": ["/lib/macOS_arm64/libnode.93.dylib"]
+        },
+        "ios": {
+            "arm64": ["/lib/iOS/*.a"]
+        },
+        "linux": {
+            "x64": ["/lib/Linux/libnode.so.93"]
+        }
+    }
+}
