+++ unity/native_src_il2cpp/cmake/quickjs/backend.rc
@@ -1,0 +1,4 @@
+BACKEND_ARTIFACT_NAME=qjs_v8_bin
+BACKEND_ARTIFACT_WORKFLOW=build_quickjs.yml
+BACKEND_ARTIFACT_REPO=puerts/backend-quickjs
+BACKEND_URL=https://github.com/puerts/backend-quickjs/releases/download/QJS_1027/qjs_v8_bin.tgz
