+++ unity/native_src_il2cpp/cmake/v8_9.4/backend.rc
@@ -1,0 +1,4 @@
+BACKEND_ARTIFACT_NAME=v8_bin
+BACKEND_ARTIFACT_WORKFLOW=build_9.4.yml
+BACKEND_ARTIFACT_REPO=puerts/backend-v8
+BACKEND_URL=https://github.com/puerts/backend-v8/releases/download/V8_9.4.146.24_0817/v8_bin_9.4.146.24.tgz
